LayoutTests/platform/ios/fast/dom/Element/scroll-width-visible-expected.txt

Tests that element.clientWidth of visible element with scrollbars returns proper values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL document.getElementById('scrollDiv').clientWidth should be 190. Was 200.
PASS successfullyParsed is true

TEST COMPLETE
